Only enabled in nodejs-ko

+++ scripts/deploy.js
@@ -7,10 +7,16 @@ const { execSync } = require("child_process")
 module.exports = (robot) => {
   // Command: ping!
   robot.respond(/ping!/, (res) => {
+    if (res.message.room !== "54e9664d15522ed4b3dc4eb3") {
+      return
+    }
     res.send("pong!")
   })
   // Command: deploy
   robot.respond(/deploy$/, (res) => {
+    if (res.message.room !== "54e9664d15522ed4b3dc4eb3") {
+      return
+    }
     const commands = [
       "rm -rf nodejs-ko",
       "git clone --depth 1 --branch master https://github.com/nodejs/nodejs-ko.git nodejs-ko",
